// prisma/generated/type-graphql/resolvers/outputs/ProductsGroupBy.d.ts
import { ProductsAvgAggregate } from "../outputs/ProductsAvgAggregate";
import { ProductsCountAggregate } from "../outputs/ProductsCountAggregate";
import { ProductsMaxAggregate } from "../outputs/ProductsMaxAggregate";
import { ProductsMinAggregate } from "../outputs/ProductsMinAggregate";
import { ProductsSumAggregate } from "../outputs/ProductsSumAggregate";
export declare class ProductsGroupBy {
    id: bigint;
    sku: bigint;
    thumbnail: string | null;
    name: string;
    price: string | null;
    quantity_per_source: string | null;
    salable_quantity: string | null;
    visibility: string | null;
    status: boolean | null;
    websites: string | null;
    product: string | null;
    tax_class: string | null;
    stock_status: string | null;
    weight: string | null;
    categories: string | null;
    created_at: Date | null;
    country: string | null;
    size: string | null;
    format: string | null;
    height: string | null;
    content: string | null;
    short_description: string | null;
    image: string | null;
    meta_title: string | null;
    meta_keywords: string | null;
    meta_description: string | null;
    meta_url: string | null;
    file: string | null;
    manufacture: string | null;
    attributes: string | null;
    brand: string | null;
    contract: string | null;
    cost_string: string | null;
    customer_type: string | null;
    family: string | null;
    manufacturer_part_number: string | null;
    occassions: string | null;
    part_number: string | null;
    related_product: bigint | null;
    tags: string | null;
    types: string | null;
    variants: string | null;
    zone: string | null;
    brands: string | null;
    coupons: string | null;
    credit_memos: string | null;
    manufacturer: string | null;
    mediamanager: string | null;
    occassions_occassionsToproducts: string | null;
    orders: string | null;
    other_products: string | null;
    product_attribute: string | null;
    product_types: string | null;
    products: string | null;
    quotes: string | null;
    rating: string | null;
    returns: string | null;
    tax_rate: string | null;
    tax_rule: string | null;
    trainings: string | null;
    workspaces: string | null;
    _count: ProductsCountAggregate | null;
    _avg: ProductsAvgAggregate | null;
    _sum: ProductsSumAggregate | null;
    _min: ProductsMinAggregate | null;
    _max: ProductsMaxAggregate | null;
}